requirements.txt
fastapi
uvicorn
openai
instructor
tiktoken
python-dotenv
pydantic